barebones booking

+++ src/main/java/htl/steyr/rdp/repository/CustomerRepository.java
@@ -1,7 +1,8 @@
 package htl.steyr.rdp.repository;
 
 import htl.steyr.rdp.model.Customer;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
